+++ app/Http/Controllers/productController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\category;
 use App\Models\product;
+use App\Models\Cart;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 use PHPUnit\Framework\Attributes\RequiresSetting;
@@ -12,12 +13,12 @@ class productController extends Controller
 {
     public function index(){
         $result= product::all();
-        return view("product.index",["products"=> $result]);
+        return view("dashboard.products.index",["products"=> $result]);
     }
 
     public function create(){
         $categories=Category::all();
-        return view("product.createProduct",["categories"=>$categories]);
+        return view("dashboard.products.create",["categories"=>$categories]);
     }
     public function store(){
         
@@ -63,10 +64,16 @@ class productController extends Controller
         
         return view("singleProduct",["product"=> $product,"related"=> $relatedProducts,"category"=> $categoryName]);
     }
+    public function showForAdmin($productId){
+
+        $product= product::findOrFail($productId );
+        $categoryName=category::where("id",$product->category_id)->value('name');
+        return view("dashboard.products.show",["product"=> $product,"categoryName"=> $categoryName]);
+    }
     public function edit(product $product){
         
         $categories=category::all();
-        return view("product.editProduct",["product"=> $product,'categories'=> $categories]);
+        return view("dashboard.products.edit",["product"=> $product,'categories'=> $categories]);
 
     }
 
@@ -77,35 +84,53 @@ class productController extends Controller
             'price'=> ['required','gt:0'],
             'quantity'=> ['required','gt:0'],
             'category_id'=> ['required','exists:categories,id'],
-            'imagepath'=> ['required','mimes:jpeg,jpg,png,gif'],
+            'imagepath'=> ['mimes:jpeg,jpg,png,gif'],
         ]);
+        $isImagePath=0;
+        $imagepath="";
         if(request()->has('imagepath')){
+            
             $file=request()->file('imagepath');
             $path=Storage::disk('public')->put('products',$file);
+            $isImagePath=1;
+            $imagepath=$path;
         }
         $name=request()->name;
         $description=request()->description;
         $price=request()->price;
         $quantity=request()->quantity;
         $category_id=request()->category_id;
-        $imagepath=$path;
         
         $singleProduct=product::findOrFail($productId);
-        Storage::disk('public')->delete($singleProduct->imagepath);
-        $singleProduct->update([
-            'name'=> $name,
-            'price'=> $price,
-            'quantity'=> $quantity,
-            'description'=>$description,
-            'imagepath'=>$imagepath,
-            'category_id'=>$category_id,
-        ]);
+        if($isImagePath== 1){
+
+            Storage::disk('public')->delete($singleProduct->imagepath);
+            $singleProduct->update([
+                'name'=> $name,
+                'price'=> $price,
+                'quantity'=> $quantity,
+                'description'=>$description,
+                'imagepath'=>$imagepath,
+                'category_id'=>$category_id,
+            ]);
+        }else{
+            $singleProduct->update([
+                'name'=> $name,
+                'price'=> $price,
+                'quantity'=> $quantity,
+                'description'=>$description,
+                'category_id'=>$category_id,
+            ]);
+        }
         
         
         return to_route('product.index');
     }
     public function destroy($productId){
         $product=product::findOrFail($productId);
+        if(count(Cart::where("product_id",$productId)->get())){
+            Cart::where("product_id",$productId)->delete();
+        }
         Storage::disk('public')->delete( $product->imagepath);
         $product->delete();
